Fix an incorrect heading in README.rst
The heading was using a single equal mark whereas the rest of the
documentation was using a two line equal mark to delimit the heading.
The later caused an incorrect parsing in PyPI documentation. Since PyPI
don't accept files that were uploaded before a new version is added,
although not significant changes were made.